pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "polarsclaw"
version = "0.1.0"
description = "A local-first, single-user personal AI assistant — OpenClaw-inspired, built on LangChain DeepAgents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    # Agent framework
    "langgraph>=0.2.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.3.0",

    # Web framework (Gateway)
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=13.0",

    # Data & storage
    "aiosqlite>=0.20.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # Scheduling
    "apscheduler>=3.10.0,<4.0.0",
    "croniter>=2.0.0",

    # CLI
    "click>=8.1.0",
    "rich>=13.0.0",

    # Utilities
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "watchfiles>=0.22.0",
]

[project.optional-dependencies]
deepagents = ["deepagents>=0.5.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
polarsclaw = "polarsclaw.cli.main:cli"

[project.entry-points."polarsclaw.plugins"]
# Example: myplugin = "myplugin:register"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["polarsclaw"]
